requirements.txt
# PCBA Thermal Mapper Dependencies
# Core dependencies
numpy>=1.21.0
pandas>=1.5.0
Pillow>=9.0.0
openpyxl>=3.0.0

# Computer Vision
opencv-python>=4.5.0

# Deep Learning - YOLOv8
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
